+++ pages/Homepage/homepage.js
@@ -1,0 +1,33 @@
+import Head from 'next/head';
+import MetaDefaults from '../../components/MetaDefaults';
+import Loader from '../../components/Loader';
+
+const Homepage = ({ homepage }) => {
+  if (homepage) {
+    return (
+      <div>
+        <Head>
+          <title>Website Demo Homepage | Website Demo</title>
+          <meta
+            name='description'
+            content='Welcome to this website demo. Please explore and test out the functionality!'
+          />
+          <meta property='og:title' content='Homepage' />
+          <meta
+            property='og:description'
+            content='Welcome to this website demo. Please explore and test out the functionality!'
+          />
+        </Head>
+        <MetaDefaults />
+        <section className='section'>
+          <div className='container'>
+            <h1>{homepage.header}</h1>
+          </div>
+        </section>
+      </div>
+    );
+  }
+  return <Loader />;
+};
+
+export default Homepage;