+++ utils/config.js
@@ -1,0 +1,4 @@
+const defaults = {
+  // contentful amount in seconds after which a page re-generation can occur, used in getStaticProps
+  pageRevalidate: 60,
+};